requirements.txt
cycler>=0.11.0
filterpy>=1.4.5
fonttools>=4.38.0
kiwisolver>=1.4.5
matplotlib>=3.5.3
numpy>=1.21.6
opencv-contrib-python>=4.12.0.88
opencv-python>=4.12.0.88
packaging>=24.0
Pillow>=9.5.0
pyparsing>=3.1.4
pyserial>=3.5
python-dateutil>=2.9.0.post0
scipy>=1.7.3
six>=1.17.0
typing_extensions>=4.7.1
wincertstore>=0.2
